`refactor`: changes views pkg name to routes
Renomeia o pacote de views para routes, a fim de deixá-lo mais
generalizado.
